+++ package.json
@@ -13,7 +13,7 @@
     }
   },
   "scripts": {
-    "test": "xs test",
+    "test": "xs test --silent",
     "clean": "xs clean",
     "start": "xs serve",
     "watch": "xs watch",
