+++ test/htmlaudio.test.ts
@@ -1,0 +1,3 @@
+import { suite } from './suite';
+
+suite(true);
